Refactor elements color picker
